a reject reason for vhosts

+++ pyserv.py
@@ -405,13 +405,14 @@ class Services:
 								if uid != 0:
 									self.vhost(uid)
 									self.memo(data[0])
+					elif len(arg) > 2:
 						if arg[0].lower() == "reject":
 							for data in self.query("select * from vhosts where active = '0' and user = '%s'" % arg[1]):
 								self.query("delete from vhosts where user = '%s'" % str(data[0]))
 								self.omsg(source, "vHost for user\2 %s\2 has been rejected" % str(data[0]))
-								self.query("insert into memo values ('%s', 'Q', 'Your vHost\2 %s\2 has been rejected.')" % (data[0], data[1]))
+								self.query("insert into memo values ('%s', 'Q', 'Your vHost\2 %s\2 has been rejected. Reason: %s')" % (data[0], data[1], _mysql.escape_string(' '.join(arg[2:]))))
 								self.memo(data[0])
-					else: self.omsg(source, "Syntax: VHOST <list>/<activate>/<reject> [<user>]")
+					else: self.omsg(source, "Syntax: VHOST <list>/<activate>/<reject> [<user>] [reason]")
 				elif cmd == "global":
 					self.omsg("$*", "[%s] " % self.nick(source) + args)
 				elif cmd == "feedback":
